trezorlib: shuffling things from client
