+++ trezorlib/tests/device_tests/test_msg_signtx_segwit.py
@@ -22,9 +22,8 @@ from .conftest import TREZOR_VERSION
 from .common import TrezorTest
 
 from trezorlib import messages as proto
-from trezorlib.client import CallException
-from trezorlib.tools import parse_path
 from trezorlib.tx_api import TxApiInsight
+from trezorlib.tools import parse_path, CallException
 
 TxApiTestnet = TxApiInsight("insight_testnet")
