+++ trezorlib/tests/device_tests/test_op_return.py
@@ -20,8 +20,7 @@ import pytest
 from .common import TrezorTest
 from .conftest import TREZOR_VERSION
 from trezorlib import messages as proto
-from trezorlib.client import CallException
-
+from trezorlib.tools import CallException
 
 TXHASH_d5f65e = unhexlify('d5f65ee80147b4bcc70b75e4bbf2d7382021b871bd8867ef8fa525ef50864882')
